+++ hart/__main__.py
@@ -213,7 +213,8 @@ class HartCLI:
 
 
     def cli_create_minion_from_role(self, args):
-        kwargs = get_minion_arguments_for_role(args.config, args.role, args.provider, args.region)
+        kwargs = get_minion_arguments_for_role(
+            args.config, args.role, args.provider, args.region, vars(args))
         for key, val in kwargs.items():
             setattr(args, key, val)
         self.cli_create_minion(args)