+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 
+UNRELEASED -
+-------------------
+
+## Fixed
+- Merge cli parameters with parameters from role parameters in the config file.
+
+
 0.14.0 - 2021-04-26
 -------------------